modified debug message

+++ src/board/v3/nap/track_channel.c
@@ -516,12 +516,12 @@ void nap_track_read_results(u8 channel,
     s->sw_carr_phase += phase_incr;
     hw_carr_phase = ((s64)t->CARR_PHASE_INT << 32) | t->CARR_PHASE_FRAC;
     if (s->sw_carr_phase != hw_carr_phase) {
-      log_error_mesid(s->mesid, "%12llu reckon err SW %+.6lf  HW %+.6lf DIFF %+.6f",
+      log_error_mesid(s->mesid, "%12llu reckon err SW %+.9lf  HW %+.9lf DIFF %+.9lf",
         s->reckon_counter,
-        s->sw_carr_phase / NAP_TRACK_CARRIER_PHASE_UNITS_PER_CYCLE,
-        hw_carr_phase / NAP_TRACK_CARRIER_PHASE_UNITS_PER_CYCLE,
-        (s->sw_carr_phase / NAP_TRACK_CARRIER_PHASE_UNITS_PER_CYCLE) -
-        (hw_carr_phase / NAP_TRACK_CARRIER_PHASE_UNITS_PER_CYCLE));
+        (double) s->sw_carr_phase / NAP_TRACK_CARRIER_PHASE_UNITS_PER_CYCLE,
+        (double) hw_carr_phase / NAP_TRACK_CARRIER_PHASE_UNITS_PER_CYCLE,
+        (double) (s->sw_carr_phase / NAP_TRACK_CARRIER_PHASE_UNITS_PER_CYCLE) -
+        (double) (hw_carr_phase / NAP_TRACK_CARRIER_PHASE_UNITS_PER_CYCLE));
     }
 #endif /* PIKSI_RELEASE */
   }
